requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
aiosqlite>=0.20.0
aiohttp>=3.10.0
pydantic>=2.0
pyyaml>=6.0
dotenv